api/backboard-give-brain/pyproject.toml
[tool.poetry]
name = "backboard-give-brain"
version = "0.1.0"
description = ""
authors = ["zoelinsg <zoelin.sg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
backboard-sdk = "^1.5.12"
python-dotenv = "^1.2.2"
requests = "^2.33.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
